refactor(eval-workflows): 归位 Claude Runtime Adapter
