+++ src/eval-workflows/runtime-adapter/adapters/claude/resources.ts
@@ -11,19 +11,19 @@ import {
   resolveEffectiveExecutionControl,
   type EvaluationDefinition,
   type JsonValue,
-} from '../../../evaluation-core/contracts/index.js';
+} from '../../../../evaluation-core/contracts/index.js';
 import {
   ExecutionPortFailure,
   type ExecutionContent,
   type ExecutorTrialContext,
-} from '../../../evaluation-core/execution/index.js';
-import type { CliMockHandle } from '../../../executors/mock-runtime/runtime.js';
-import type { Mock, MockMatch, MockReturn } from '../../../inputs/contracts/mock.js';
-import type { RuntimeBindingOf } from '../types.js';
+} from '../../../../evaluation-core/execution/index.js';
+import type { CliMockHandle } from '../../../../executors/mock-runtime/runtime.js';
+import type { Mock, MockMatch, MockReturn } from '../../../../inputs/contracts/mock.js';
+import type { RuntimeBindingOf } from '../../types.js';
 import type {
   OmkBindingResourceLease,
   OmkLeasedHostResource,
-} from '../resource-leases/types.js';
+} from '../../resource-leases/types.js';
 
 const DescriptorSchema = z.object({
   resourceId: z.string().min(1),
