+++ src/eval-workflows/runtime-adapter/adapters/index.ts
@@ -1,6 +1,6 @@
 export * from './anthropic/api.js';
-export * from './claude-cli.js';
-export * from './claude-sdk.js';
+export * from './claude/cli.js';
+export * from './claude/sdk.js';
 export * from './codex-cli.js';
 export * from './codex-sdk.js';
 export * from './custom/command.js';
