+++ test/eval-workflows/runtime-adapter/adapters/claude/sdk.test.ts
@@ -14,15 +14,15 @@ import {
   type JsonValue,
   type SchemaIdentity,
   type Sha256Digest,
-} from '../../../src/evaluation-core/contracts/index.js';
-import { prepareEvaluationPlan } from '../../../src/evaluation-core/compiler/index.js';
+} from '../../../../../src/evaluation-core/contracts/index.js';
+import { prepareEvaluationPlan } from '../../../../../src/evaluation-core/compiler/index.js';
 import {
   InMemoryRuntimeEventSequencer,
   executeRunPlan,
   type ExecutionExecutor,
   type ExecutorAttemptResult,
   type ExecutorTrialContext,
-} from '../../../src/evaluation-core/execution/index.js';
+} from '../../../../../src/evaluation-core/execution/index.js';
 import {
   createClaudeSdkCoreSchemaValidators,
   createClaudeSdkExecutorAdapter,
@@ -35,12 +35,12 @@ import {
   type OmkLeasedHostResource,
   type ResolvedClaudeSdkRuntime,
   type RuntimeBindingOf,
-} from '../../../src/eval-workflows/runtime-adapter/index.js';
+} from '../../../../../src/eval-workflows/runtime-adapter/index.js';
 import {
   testRuntime,
   validDefinition,
   validPolicy,
-} from '../../evaluation-core/compiler/fixtures.js';
+} from '../../../../evaluation-core/compiler/fixtures.js';
 
 const roots = new Set<string>();
